Move the server to its own package
