--- log.go
@@ -1,7 +1,0 @@
-package main
-
-import "go.uber.org/zap"
-
-func logger() *zap.SugaredLogger {
-	return zap.S()
-}
